+++ code/isolation.cpp
@@ -156,7 +156,8 @@ int main(int argc, char **argv) {
 
       IsolationTask *task = new IsolationTask(cache, output_directory, bounds, minIsolation);
       results.push_back(threadPool->enqueue([=] {
-            return task->run(lat, lng, peakbagger_peaks);
+            return task->run(static_cast<float>(lat), static_cast<float>(lng),
+                             peakbagger_peaks);
           }));
     }
   }
